test: move system tests
