+++ crates/tinymist-world/src/font/system.rs
@@ -262,3 +262,32 @@ impl Default for SystemFontSearcher {
         Self::new()
     }
 }
+
+#[cfg(test)]
+mod tests {
+    use std::sync::Arc;
+
+    use super::*;
+
+    #[test]
+    fn edit_fonts() {
+        use clap::Parser as _;
+
+        use crate::args::CompileOnceArgs;
+
+        let args = CompileOnceArgs::parse_from(["tinymist", "main.typ"]);
+        let mut verse = args
+            .resolve_system()
+            .expect("failed to resolve system universe");
+
+        let fonts: Vec<_> = verse.font_resolver.get_fonts().collect();
+
+        let new_resolver = FontResolverImpl::new_with_fonts(
+            vec![],
+            fonts
+                .into_iter()
+                .map(|(info, slot)| (info.clone(), slot.clone())),
+        );
+        verse.increment_revision(|verse| verse.set_fonts(Arc::new(new_resolver)));
+    }
+}
